+++ src/test/converter2/behavior/inheritDocWarnings.ts
@@ -4,6 +4,8 @@
  */
 export const source = 123;
 
+export const source2 = 123;
+
 /**
  * Target summary
  * @inheritDoc source
@@ -15,3 +17,13 @@ export const target1 = 123;
  * @remarks Target remarks
  */
 export const target2 = 123;
+
+/**
+ * @inheritDoc doesNotExist
+ */
+export const target3 = 123;
+
+/**
+ * @inheritDoc source2
+ */
+export const target4 = 123;